Move read_midi to utils.py
