+++ train_model.py
@@ -17,6 +17,8 @@ import sys
 from pebble import ProcessPool
 from concurrent.futures import TimeoutError
 
+from utils import *
+
 path = "/Users/jscherman/GolandProjects/music-generator/gwern/midis/Big_Data_Set"
 frequent_notes_threshold = 256
 
@@ -28,108 +30,6 @@ kernel_size = 3
 epochs = 50
 
 len_of_predictions = 30
-
-def read_midi_wrapper(args):
-    return read_midi(*args)
-
-def read_midi(file, current, total):
-    print(f'reading: {str(current)}/{str(total)}', flush=True)
-    notes=[]
-    notes_to_parse = None
-
-    #parsing a midi file
-    try:
-        midi = converter.parse(file)
-    except:
-        return np.array([])
-
-    #grouping based on different instruments
-    s2 = instrument.partitionByInstrument(midi)
-    if not s2:
-        return np.array([])
-
-    monophonics = [
-        'Flute',
-        'Baritone Saxophone',
-        'Trombone',
-        'Tenor Saxophone',
-        'Clarinet',
-        'Alto Saxophone',
-        'Bass',
-        'Voice',
-        'Acoustic Bass',
-        'Trumpet',
-        'Oboe']
-
-    #Looping over all the instruments
-    for part in s2.parts:
-
-        #select elements of only piano
-        # if 'Piano' in str(part):
-        if any(x in str(part) for x in monophonics):
-
-            # Transpose to C major / A minor
-            try:
-                key = part.analyze('key')
-                tonic = key.getTonic()
-            except:
-                return np.array([])
-
-            if key.mode == 'major':
-                part.transpose(tonic.pitchClass * -1, inPlace=True)
-            else:
-                part.transpose((tonic.pitchClass * -1) - 3, inPlace=True)
-
-            notes_to_parse = part.recurse()
-
-            #finding whether a particular element is note or a chord
-            for element in notes_to_parse[:1024]:
-#                 print(f'notes to parse: {len(list(notes_to_parse))}')
-                #note
-                if isinstance(element, note.Note):
-                    notes.append(str(element.pitch))
-
-                #chord
-                elif isinstance(element, chord.Chord):
-                    notes.append('.'.join(str(n) for n in element.normalOrder))
-    print(f'finish reading: {str(current)}/{str(total)}', flush=True)
-    return np.array(notes)
-
-def convert_to_midi(prediction_output, filename):
-
-    offset = 0
-    output_notes = []
-
-    # create note and chord objects based on the values generated by the model
-    for pattern in prediction_output:
-
-        # pattern is a chord
-        if ('.' in pattern) or pattern.isdigit():
-            notes_in_chord = pattern.split('.')
-            notes = []
-            for current_note in notes_in_chord:
-
-                cn=int(current_note)
-                new_note = note.Note(cn)
-                new_note.storedInstrument = instrument.Piano()
-                notes.append(new_note)
-
-            new_chord = chord.Chord(notes)
-            new_chord.offset = offset
-            output_notes.append(new_chord)
-
-        # pattern is a note
-        else:
-
-            new_note = note.Note(pattern)
-            new_note.offset = offset
-            new_note.storedInstrument = instrument.Piano()
-            output_notes.append(new_note)
-
-        # increase offset each iteration so that notes do not stack
-        offset += 1
-    midi_stream = stream.Stream(output_notes)
-    midi_stream.write('midi', fp=filename)
 
 from glob import glob
 from multiprocessing.dummy import Pool as ThreadPool
